+++ maths/add.go
@@ -17,7 +17,7 @@ func Add(number1, number2 interface{}) interface{} {
 		panic("'number2' is not a number")
 	}
 
-	outputType := getOutputNumberType(number1, number2)
+	outputType := internal.GetOutputNumberType(number1, number2)
 
 	number1Value := reflect.ValueOf(number1)
 	number2Value := reflect.ValueOf(number2)
@@ -42,49 +42,4 @@ func Add(number1, number2 interface{}) interface{} {
 	}
 
 	return output.Interface()
-}
-
-// Returns a type which has higher rank in between two passed variables.
-func getOutputNumberType(number1, number2 interface{}) reflect.Type {
-	numbe1Type := reflect.TypeOf(number1)
-	numbe2Type := reflect.TypeOf(number2)
-
-	number1Rank := getNumberTypeRank(numbe1Type.Kind())
-	number2Rank := getNumberTypeRank(numbe2Type.Kind())
-
-	if number1Rank >= number2Rank {
-		return numbe1Type
-	} else {
-		return numbe2Type
-	}
-}
-
-// Returns rank of the passed king
-func getNumberTypeRank(kind reflect.Kind) int {
-	switch kind {
-	case reflect.Int8:
-		return 0
-	case reflect.Uint8:
-		return 1
-	case reflect.Int16:
-		return 2
-	case reflect.Uint16:
-		return 3
-	case reflect.Int32, reflect.Int:
-		return 4
-	case reflect.Uint32:
-		return 5
-	case reflect.Int64:
-		return 6
-	case reflect.Uint64:
-		return 7
-	case reflect.Uintptr:
-		return 8
-	case reflect.Float32:
-		return 9
-	case reflect.Float64:
-		return 10
-	default:
-		return -1
-	}
 }